pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "bigbang-py"
version = "0.4.4"
authors = [
  { name="BigBang Team", email="bigbang-dev@data-activism.net" },
]
description = "BigBang is a toolkit for studying communications data from collaborative projects. It currently supports analyzing mailing lists from Sourceforge, Mailman, ListServ, or .mbox files."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "networkx",
    "requests",
    "pandas",
    "beautifulsoup4",
    "chardet",
    "html2text",
    "python-dateutil",
    "pytz",
    "tqdm",
    "pyyaml",
    "certifi",
    "levenshtein",
    "ietfdata",
    "python-docx",
    "GitPython",
    "nbformat",
    "nbconvert",
    "validator-collection",
]

[project.optional-dependencies]
test = ['pytest', 'coverage', 'testfixtures']
dev = ['bigbang-py[test]', 'black', 'isort', 'pre-commit']

[project.urls]
"Homepage" = "https://github.com/datactive/bigbang"
"Bug Tracker" = "https://github.com/datactive/bigbang/issues"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
include_trailing_comma = "True"
known_third_party = "celery,django,environ,pyquery,pytz,redis,requests,rest_framework"

[tool.ruff]
ignore = ['F401', 'E501', 'E731', 'F403', 'F821', 'F541', 'E402', 'F405']
